pyproject.toml
[project]
name = "programming"
version = "0.1.0"
description = ""
authors = [
    {name = "Elizabeth",email = "rxxivkey@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
flake8 = "^7.2.0"
isort = "^6.0.1"
mypy = "^1.15.0"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^6.1.1"

[tool.isort]
# максимальная длина строки
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'
